Moved GatherLocalsVisitor to its own file
